refactor: Move all shared drive hook in same folder
